decouple Transform from REST ruleset
