+++ java/core/src/main/java/co/worklytics/psoxy/rules/Rules2.java
@@ -3,10 +3,7 @@ package co.worklytics.psoxy.rules;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import lombok.*;
-import lombok.experimental.SuperBuilder;
 import lombok.extern.java.Log;
 
 import java.io.Serializable;
@@ -42,6 +39,7 @@ public class Rules2 implements RuleSet, Serializable {
     @Builder.Default
     Boolean allowAllEndpoints = false;
 
+
     /**
      * add endpoints from other ruleset to this one
      * @param endpointsToAdd to be added
@@ -53,13 +51,13 @@ public class Rules2 implements RuleSet, Serializable {
         return builder.build();
     }
 
-
     @JsonPropertyOrder(alphabetic = true)
     @Builder
     @AllArgsConstructor //for builder
     @NoArgsConstructor //for Jackson
     @Getter
     public static class Endpoint {
+
         String pathRegex;
 
         @JsonInclude(value=JsonInclude.Include.NON_EMPTY)
@@ -67,71 +65,9 @@ public class Rules2 implements RuleSet, Serializable {
         List<Transform> transforms = new ArrayList<>();
     }
 
+
     //TODO: fix YAML serialization with something like
     // https://stackoverflow.com/questions/55878770/how-to-use-jsonsubtypes-for-polymorphic-type-handling-with-jackson-yaml-mapper
 
-    @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "method")
-    @JsonSubTypes({
-        @JsonSubTypes.Type(value = Redact.class, name = "redact"),
-        @JsonSubTypes.Type(value = Pseudonymize.class, name = "pseudonymize"),
-        @JsonSubTypes.Type(value = PseudonymizeEmailHeader.class, name = "pseudonymizeEmailHeader"),
-    })
-    @SuperBuilder
-    @AllArgsConstructor //for builder
-    @NoArgsConstructor //for Jackson
-    @Getter
-    @EqualsAndHashCode(callSuper = false)
-    public static abstract class Transform {
-
-        //NOTE: this is filled for JSON, but for YAML a YAML-specific type syntax is used:
-        // !<pseudonymize>
-        // Jackson YAML can still *read* yaml-encoded transform with `method: "pseudonymize"`
-        @JsonInclude(JsonInclude.Include.NON_NULL)
-        String method;
-
-        @Singular
-        List<String> jsonPaths;
-    }
-
-
-    @NoArgsConstructor //for jackson
-    @SuperBuilder
-    @Getter
-    @EqualsAndHashCode(callSuper = true)
-    public static class Redact extends Transform {
-
-        public static Redact ofPaths(String... jsonPaths) {
-            return Redact.builder().jsonPaths(Arrays.asList(jsonPaths)).build();
-        }
-    }
-    @NoArgsConstructor //for jackson
-    @SuperBuilder
-    @Getter
-    public static class PseudonymizeEmailHeader extends Transform {
-
-        public static PseudonymizeEmailHeader ofPaths(String... jsonPaths) {
-            return PseudonymizeEmailHeader.builder().jsonPaths(Arrays.asList(jsonPaths)).build();
-        }
-    }
-
-
-
-    @SuperBuilder
-    @AllArgsConstructor //for builder
-    @NoArgsConstructor //for Jackson
-    @Getter
-    public static class Pseudonymize extends Transform {
-
-        @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-        @Builder.Default
-        Boolean includeOriginal = false;
-
-        //TODO: support this somehow ...
-        //String defaultScopeId;
-
-        public static Pseudonymize ofPaths(String... jsonPaths) {
-            return Pseudonymize.builder().jsonPaths(Arrays.asList(jsonPaths)).build();
-        }
-    }
 
 }
